// src/app/model/user-data.ts
export class UserData {
  result: string;
  id: string;
  password: string;
  role: string;
  name: string;
  nickname: string;
  mobile: string;
  email: string;
  sns: string;
  introduce: string;
  social_type: string;

  constructor(){
    this.result="";
    this.id="";
    this.password="";
    this.name="";
    this.role="";
    this.nickname="";
    this.mobile="";
    this.email="";
    this.sns="";
    this.introduce="";
    this.social_type="";
  }
}